% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aswb_extrapolate.R
\name{aswb_extrapolate}
\alias{aswb_extrapolate}
\title{One-shot extrapolation: PEM(INLA) + anchored Gompertz + monotone spline weight}
\usage{
aswb_extrapolate(
  df,
  t_obs,
  t_max,
  interval = 0.5,
  nsim_inla = 1000,
  nsim_ext = 1000,
  inla_threads = 1,
  anchor_t = t_max,
  anchor_mean_Sa = 0.1,
  anchor_ess = 50,
  anchor_type = c("beta", "fixed"),
  loggamma_mean = log(0.08),
  loggamma_sd = 0.5,
  scam_k = 15,
  tau_init = 1,
  tau_max = 100,
  tau_step = 1.2,
  tol = -1e-10,
  seed = NULL
)
}
\description{
One-shot extrapolation: PEM(INLA) + anchored Gompertz + monotone spline weight
}
