% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{compute_metrics}
\alias{compute_metrics}
\title{Error metrics for extrapolated survival curves}
\usage{
compute_metrics(times, S_true, S_model, t0, eval_times = c(15, 20))
}
\arguments{
\item{times}{grid}

\item{S_true}{true survival}

\item{S_model}{model survival}

\item{t0}{observation horizon}

\item{eval_times}{vector of evaluation times (default 15, 20)}
}
\value{
list of metrics
}
\description{
Error metrics for extrapolated survival curves
}
