% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_gompertz.R
\name{draw_ext_gompertz}
\alias{draw_ext_gompertz}
\title{Draw external Gompertz survival curves given an anchor constraint}
\usage{
draw_ext_gompertz(
  times,
  t_a,
  S_a_mean,
  ess_Sa = 50,
  loggamma_mean,
  loggamma_sd,
  nsim = 2000,
  anchor_type = c("beta", "fixed")
)
}
\arguments{
\item{times}{grid}

\item{t_a}{anchor time}

\item{S_a_mean}{anchor survival mean S(t_a); if anchor_type="fixed" this is exact S(t_a)}

\item{ess_Sa}{equivalent sample size for Beta prior on S(t_a) (ignored if fixed)}

\item{loggamma_mean, loggamma_sd}{log-normal prior for Gompertz gamma}

\item{anchor_type}{"beta" (default) or "fixed"}
}
\description{
Draw external Gompertz survival curves given an anchor constraint
}
