% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{rmst_numeric}
\alias{rmst_numeric}
\title{Numerical RMST on \link{t0, t*} by trapezoidal rule}
\usage{
rmst_numeric(times, S, t0, t_star)
}
\arguments{
\item{times}{grid}

\item{S}{survival on grid}

\item{t0}{lower limit}

\item{t_star}{upper limit}
}
\value{
scalar
}
\description{
Numerical RMST on \link{t0, t*} by trapezoidal rule
}
