% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{survblendr_example_data_path}
\alias{survblendr_example_data_path}
\title{Path to the packaged example CSV}
\usage{
survblendr_example_data_path()
}
\value{
A file path to \code{inst/extdata/survblendr_demo.csv}.
}
\description{
Path to the packaged example CSV
}
